praas_python/praas_py_client/requirements.txt
# © 2024 Nokia
# Licensed under the BSD 3-Clause Clear License
# SPDX-License-Identifier: BSD-3-Clause-Clear

requests
azure-security-attestation
azure-identity